requirements.txt
PyJWT>=2.3.0
werkzeug>=1.0.0,<2.1.0
autobahn>=22.7.1
alembic>=0.9.9
bcrypt>=3.1.4
colorama>=0.3.9
click>=8.0.1
flask>=1.1,<=2.1.3
Flask-SQLAlchemy>=2.3.1,<=2.5.1
flask-classful>=0.14,<0.16
Flask-CORS>=3.0.0
email_validator
validators
Flask-WTF>=0.15.1,<=1.0.1
WTForms>=2.1
flask-login>=0.5.0,<0.6.0
Flask-Security-Too>=4.0.0,<5.0.0
bleach>=4.1.0,<5.0.0
marshmallow>=3.19.0,<4.0.0
Pillow>=10.2.0
psycopg2
pgcli
cryptography>=40.0.1
pyopenssl>=23.1.1
python-dateutil>=2.6.0
requests>=2.18.4
pyasn1
service_identity>=17.0.0
SQLAlchemy>=1.2.0,<1.4.0
tqdm>=4.15.0
webargs>=7.0.0
marshmallow-sqlalchemy>=0.26.1,<0.28.1
filteralchemy-fork
filedepot>=0.5.0,<0.11.0
nplusone>=0.8.1
syslog-rfc5424-formatter>=1.1.1
simplekv>=0.13.0
Flask-KVSession-fork>=0.6.4
distro>=1.4.0
faraday-plugins>=1.24.1,<2.0.0
apispec>=6.3.0
apispec-webframeworks<=0.5.2
pyyaml
bidict==0.22.0
Flask-SocketIO>=5.0.1
pyotp>=2.6.0
Flask-Limiter>=1.3.1,<1.4.0
Flask-Mail
faraday-agent-parameters-types>=1.7.3,<1.8.0
cvss>=3.1
celery==5.4.0
kombu==5.4.2
gevent>=22.10.2
psycogreen>=1.0.2
Flask-Celery-Helper>=1.1.0
redis>=4.3.4
gevent-websocket>=0.10.1
sh
elasticsearch>=7.16.3
croniter>=0.3.30
